% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate}
\alias{annotate}
\title{Annotate point with rasters based on matching dates assoi}
\usage{
annotate(datedOccs, env, envDates, dateScale)
}
\arguments{
\item{datedOccs}{a `SpatialPointsDataFrame` where one column is labeled `date` and has class `POSIXct`, e.g., as obtained from using `lubridate::parse_date_time`}

\item{env}{a raster stack}

\item{envDates}{a vector of dates the same length as `env`. The vector should have class `POSIXct`, e.g., as obtained from using `lubridate::parse_date_time`}

\item{dateScale}{string: 'year', 'month', or 'day'}
}
\description{
Annotate point with rasters based on matching dates associated with points to dates associated with raster. Specifically, we're thinking of the points as species observations and the rasters as remotely sensed environmental layers, but they can represent any points and rasters with dates.
}
\details{
See Examples.
}
\note{
To apply multiple masks, e.g., elevation and forest cover, use separate calls to maskRS.
}
\author{
Cory Merow <cory.merow@gmail.com>,
}
