% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialUtils.R
\name{cropResampleTrim}
\alias{cropResampleTrim}
\title{Line two rasters or stacks or lists of rasters}
\usage{
cropResampleTrim(expertMap, maskListRaw)
}
\arguments{
\item{expertMap}{A binary map, either as a polygon or a raster.}

\item{maskListRaw}{A list of rasters, each corresponding to layers from which masks will eventually be made (in another function).}
}
\description{
Obtain the same extents and resample to the finest resolution layer.
}
\details{
See Examples.
}
\examples{
\dontrun{
# please see complete example in the maskRangeR_tutorial vignette
expertMap <- readOGR(paste0(dataDir,'IUCNshape/data_0.shp'))
maskListRaw <- list(treeCover = 
raster(paste0(dataDir,'/Hansen_percent_treecover.tif')), 
dem = raster(paste0(dataDir,'/DEM.tif')),
mat = raster(paste0(dataDir,'/MATsingapore.tif')),
pH = raster(paste0(dataDir,'/PHIHOX_M_sl1_250m.tif')))
# define the limits for each mask
maskBounds <- read.csv(paste0(dataDir,'/crabInfo.csv'))
# rename the limits to match the raster names in maskListRaw
maskBounds$Layer <- c('water','treeCover','dem','pH','mat')
# make sure the units match between the layers and the bounds
maskListRaw$mat <- maskListRaw$mat/10
maskListRaw$pH <- maskListRaw$pH/10
crt <- maskRangeR::cropResampleTrim(expertMap,maskListRaw)
}

}
\author{
Cory Merow <cory.merow@gmail.com>,
}
