% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDemoData.R
\name{getDemoData}
\alias{getDemoData}
\title{Downloads data to use in vignette}
\usage{
getDemoData(path)
}
\arguments{
\item{path}{character string representing the file pathway}
}
\value{
saves demonstration data to the specified pathway.
}
\description{
Canned datasets of real case studies for examples to follow along with the vignette
}
\details{
The function will download the data from a github repository as a zipped folder and will unzip it in the destination path specifies by 'path'
See Examples.
}
\note{
These data were to large to serve on CRAN, so this convenience function finds and downloads them automatically.
}
\examples{
\dontrun{
getDemoData('~/Desktop')
}

}
\author{
Peter Galante <pgalante@amnh.org>,

Cecina Babich Morrow <cbabichmorrow@amnh.org>,
}
