% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMODIS.R
\name{getModis}
\alias{getModis}
\title{download MODIS products (e.g., foret cover) based on your occurrence records}
\usage{
getModis(localSavePath, datedOccs, dateScale, dataset)
}
\arguments{
\item{localSavePath}{the pathway or directory in which to save MODIS rasters}

\item{datedOccs}{a data.frame showing longitude, latitude, and date. The date should match the date scale and of class `POSIXct`, e.g., as obtained from using `lubridate::parse_date_time`}

\item{dateScale}{string: 'year', 'month', or 'day'}

\item{dataset}{MODIS product of interest. Available are: "Percent_Tree_Cover", "Percent_Nontree_Vegetation", "Percent_Nonvegetated", "Quality", "Percent_Tree_Cover_SD", "Percent_Nonvegetated_SD", "Cloud"}
}
\value{
Saves rasters to specified path
}
\description{
Accesses yearly (2000 - 2017) MODIS products based on geographic location and date
}
\details{
See Examples.
}
\examples{
\donttest{
simulateData()
getModis(localSavePath = '', datedOccs, dateScale = "year", dataset = "Percent tree cover")
}
}
\author{
Peter Galante <pgalante@amnh.org>,
}
