% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask.R
\name{lotsOfMasks}
\alias{lotsOfMasks}
\title{Generate and apply multiple masks to a map}
\usage{
lotsOfMasks(expertRaster, maskStack, maskBounds)
}
\arguments{
\item{expertRaster}{The binary expert map (1s and 0s), rasterized with the same projection as `maskStack`}

\item{maskStack}{A stack of *named* layers from which masks will be made}

\item{maskBounds}{A data.frame with columns indicating the layer name (matching the names in maskStack), and the min and max values of that layer to be used for masking.}
}
\description{
Based on a potential distribution, environmental rasters, and bounds for suitable habitat on the environmental rasters
}
\details{
See Examples.
}
\author{
Cory Merow <cory.merow@gmail.com>,
}
