% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyMaskSensitivity.R
\name{manyMaskSensitivity}
\alias{manyMaskSensitivity}
\title{Sensitivity testing for masks}
\usage{
manyMaskSensitivity(crt, rasProj = NULL)
}
\arguments{
\item{crt}{A raster stack; the output from the function maskRangeR::cropResampleTrim}

\item{rasProj}{(optional) a character string: a proj4string showing the projection of the environmental layers. If NULL, areas will be estimated using the raster package.}
}
\value{
returns a data.frame where row names are the environmental layer name combinations, and Area is expressed in square km, unless a projection is supplied
}
\description{
Compare how masks of climate tolerances affect predicted area
}
\details{
See Examples.
}
\author{
Peter Galante <pgalante@amnh.org>
}
