% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask.R
\name{maskRanger}
\alias{maskRanger}
\title{Make a matrix of modeling decisions to be used to specify clipping rules}
\usage{
maskRanger(
  potentialDist,
  initialDist = NULL,
  maskLayers,
  logicString,
  method = "mask"
)
}
\arguments{
\item{potentialDist}{A raster stack of binary or continuous values. Supplying more than one layer will be interepreted as different time periods. Layers should follow the naming convention `Y2000`, `Y2001`, etc.}

\item{initialDist}{A raster showing a previously created optimally tuned SDM}

\item{maskLayers}{A single raster or a raster stack. If a single raster, the same mask will be applied to each layer of `potentialDist`. If a stack it must have the same number of layers as potentialDist, and each layer corresponds to a different time period.}

\item{logicString}{a character indicating the logical conditions to use for masking.}

\item{method}{A list of strings defining methods to be used, in the same order as `rsList`. 
If a single value is provided it will be applied to all rasters in `rsList`. Options include:
\itemize{
 \item{mask}{mask cells with values outside the bounds}
}}
}
\description{
Performs data driven masking of potential species distributions.
}
\details{
See Examples.
}
\note{
To apply multiple masks, e.g., elevation and forest cover, use separate calls to maskRS.
}
\examples{
r1 <- raster(nrows=108, ncols=21, xmn=0, xmx=10)
values(r1)<- runif(n = (108*21))
r2 <- raster(nrows=108, ncols=21, xmn=0, xmx=10)
values(r2)<- runif(n = (108*21))
maskRanger(initialDist = r1, maskLayers = r2, logicString = "maskLayers> 0.5")

}
\author{
Cory Merow <cory.merow@gmail.com>,
}
