% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm.R
\name{rangeSVM}
\alias{rangeSVM}
\title{Classify species ranges based on occurrence coordinates and continuous values from an SDM}
\usage{
rangeSVM(xy1, xy2, sdm = NULL, nrep = 100, weight = FALSE)
}
\arguments{
\item{xy1}{Matrix or data frame of occurrence coordinates for species 1.}

\item{xy2}{Matrix or data frame of occurrence coordinates for species 2.}

\item{sdm}{Raster or RasterStack representing environmental suitability (can be predictions from SDMs).
Default is NULL.}

\item{nrep}{Numeric for number of SVM tuning iterations. Default is 100.}

\item{weight}{Boolean. If TRUE, the species with less occurrence records is weighted higher in the SVM.
Default is FALSE.}
}
\value{
The tuned SVM model.
}
\description{
\code{rangeSVM()} returns a tuned support vector machine (SVM) model that predicts 
species identity based on predictors that are either (1) solely spatial, (2) based 
on occurrence coordinates, or (3) a combination of spatial and environmental, based 
on both occurrence coordinates and environmental suitability values. Suitability 
values can be predicted with species distribution models (SDMs; a.k.a. ecological 
niche models).
}
\details{
The tuning operation uses \code{tune.svm()} from the e1071 package, 
which performs 10-fold cross validation and selects the best model based on 
classification error. Ranges of the cost and gamma parameters are explored in 
the tuning exercise. The tuning function is iterated \code{nrep} times, and the
parameter combination used most frequently across all iterations is used to 
build a final SVM model.

When \code{sdm = NULL}, the SVM is purely spatial, based only on the occurrence 
coordinates of each species. Otherwise, the SVM is fit with both a spatial 
predictor and any additional ones added as rasters. These extra predictors can be 
based on predictions from a species distribution model (SDM; a.k.a. ecological 
niche model), and in this case would represent environmental or climatic suitability, 
depending on the variables used in the SDM.
}
\examples{
\dontrun{
# tune SVMs on coordinates only (spatial)
rangeSVM(xy1, xy2)

# tune SVMs on coordinates and SDM prediction rasters (hybrid)
rangeSVM(xy1, xy2, raster::stack(sdm1, sdm2))
}
}
