% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm.R
\name{rangeSVM_predict}
\alias{rangeSVM_predict}
\title{Generate binary raster based on predictions of SVM model.}
\usage{
rangeSVM_predict(svm, r, sdm = NULL)
}
\arguments{
\item{svm}{Model object for the SVM, returned by \code{rangeSVM()}.}

\item{r}{Raster with the extent desired for the prediction. The values for cells used for predictions must 
have non-NA values, but the particular values do not matter.}

\item{sdm}{Raster or RasterStack representing environmental suitability (can be predictions from SDMs).
These rasters must match the predictor variables used in the SVM. Default is NULL.}
}
\value{
The binary Raster representing the SVM predictions.
}
\description{
\code{rangeSVM_predict()} returns a binary raster representing the ranges of the two species
predicted by the fitted SVM tuned with \code{rangeSVM()}.
}
\details{
The values of the output raster are 1 and 2, corresponding to xy1 and xy2 for \code{rangeSVM()}.
These values represent the identities of the species.
}
\examples{
\dontrun{
# raster prediction for xy SVM (spatial)
svm.xy <- rangeSVM(xy1, xy2)
svm.xy.r <- rangeSVM_predict(svm.xy, r)
plot(svm.xy.r)

# raster prediction for xyEnv SVM (spatial + SDM predictions)
svm.xyEnv <- rangeSVM(xy1, xy2, raster::stack(sdm1, sdm2))
svm.xyEnv.r <- rangeSVM_predict(svm.xyEnv, r, raster::stack(sdm1, sdm2))
plot(svm.xyEnv.r)
}
}
