% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm.R
\name{rasterizeRangeSVM}
\alias{rasterizeRangeSVM}
\title{Generate binary raster based on predictions of SVM model.}
\usage{
rasterizeRangeSVM(svm, r, sdm = NULL)
}
\arguments{
\item{svm}{Model object for the SVM, returned by \code{rangeSVM()}.}

\item{r}{Raster with the extent desired for the prediction. The values do not matter.}

\item{sdm}{Raster or RasterStack representing environmental suitability (can be predictions from SDMs).
These rasters must match the predictor variables used in the SVM. Default is NULL.}
}
\value{
The binary Raster representing the SVM predictions.
}
\description{
\code{rasterizeRangeSVM()} returns a binary raster representing the ranges of the two species
predicted by the fitted SVM tuned with \code{rangeSVM()}.
}
\details{
The values of the output raster are 1 and 2, corresponding to xy1 and xy2 for \code{rangeSVM()}.
These values represent the identities of the species.
}
\examples{
\dontrun{
# raster prediction for spatial SVM
svm.sp <- rangeSVM(xy1, xy2)
svm.sp.r <- rasterizeRangeSVM(svm.sp, r)
plot(svm.sp.r)

# raster prediction for hybrid SVM
svm.hyb <- rangeSVM(xy1, xy2, raster::stack(sdm1, sdm2))
svm.hyb.r <- rasterizeRangeSVM(svm.hyb, r, raster::stack(sdm1, sdm2))
plot(svm.hyb.r)
}
}
