% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDrivenMask.R
\name{RSmask}
\alias{RSmask}
\title{Creating a minimum and maximum value mask to apply to models}
\usage{
RSmask(datedOccs, RSenv, dateScale, Model)
}
\arguments{
\item{datedOccs}{A .csv file containing 3 columns of longitude, latitude, date. Dates must either be year only, or YMD in a lubridate accepted format}

\item{RSenv}{A rasterStack of remotely sensed data (i.e., percent Forest Cover) with names as dates (eg. Y2000, Y2001) or date-times YMD (e.g., Y2000-02-01)}

\item{dateScale}{a character vector of the scale the RSenv and corresponding datedOccs. Could be "year", "month", "day".}

\item{Model}{A raster file (generally in continuous output) with extent, resolution, etc. matching RSenv.}
}
\description{
Function masks a model raster by the minimum and maximum observed values at occurrences. Occurrence localities with dates paired with raster layers from those same dates are used to determine the minimum and maximum observed values through time.
}
\author{
Peter Galante <pgalante@amnh.org>
}
