% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialUtils.R
\name{cropResampleTrim}
\alias{cropResampleTrim}
\title{Line two rasters or stacks or lists of rasters}
\usage{
cropResampleTrim(expertMap, maskListRaw)
}
\arguments{
\item{expertMap}{A binary map, either as a polygon or a raster.}

\item{maskListRaw}{A list of rasters, each corresponding to layers from which masks will eventually be made (in another function).}
}
\description{
Obtain the same extents and resample to the finest resolution layer.
}
\details{
See Examples.
}
\author{
Cory Merow <cory.merow@gmail.com>,
}
