% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialUtils.R
\name{cropResampleTrim}
\alias{cropResampleTrim}
\title{Line two rasters or stacks or lists of rasters}
\usage{
cropResampleTrim(expertMap, maskListRaw)
}
\arguments{
\item{expertMap}{A binary map, either as a polygon or a raster.}

\item{maskListRaw}{A list of rasters, each corresponding to layers from which masks will eventually be made (in another function).}
}
\value{
a list
}
\description{
Obtain the same extents and resample to the finest resolution layer.
}
\details{
See Examples.
}
\examples{
# please see complete example in the maskRangeR_tutorial vignette
r1 <- raster(nrows=108, ncols=21, xmn=0, xmx=10)
values(r1) <- runif((108*21))
r1[r1>0.5] <-1
r1[r1<0.5] <- NA
r1shape <- rasterToPolygons(r1)
r2 <- raster(nrows=108, ncols=21, xmn=0, xmx=10)
values(r2) <- runif((108*21))
cropResampleTrim(expertMap = r1shape, maskListRaw = list(r2))

}
\author{
Cory Merow <cory.merow@gmail.com>,
}
