% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMODIS.R
\name{getModis}
\alias{getModis}
\title{download MODIS products (e.g., foret cover) based on your occurrence records}
\usage{
getModis(localSavePath, datedOccs, dateScale, dataset)
}
\arguments{
\item{localSavePath}{the pathway or directory in which to save MODIS rasters}

\item{datedOccs}{a `SpatialPointsDataFrame` where one column is labeled `date` and has class `POSIXct`, e.g., as obtained from using `lubridate::parse_date_time`}

\item{dateScale}{string: 'year', 'month', or 'day'}

\item{dataset}{MODIS product of interest. Available are: “Percent tree cover”, “Percent nontree vegetation”, “Percent nonvegetated”, “Quality”, “Percent tree cover 50”, “Percent nonvegetated sd”}
}
\value{
Saves rastersto specified path
}
\description{
Accesses yearly (2000 - 2017) MODIS products based on geographic location and date
}
\details{
See Examples.
}
\examples{
\donttest{
simulateData()
getModis(localSavePath = '', datedOccs, dateScale = "year", dataset = "Percent tree cover")
}
}
\author{
Peter Galante <pgalante@amnh.org>,
}
