% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm.R
\name{rangeSVM}
\alias{rangeSVM}
\title{Classify species ranges based on occurrence coordinates and SDM scores.}
\usage{
rangeSVM(xy1, xy2, ..., sdm = NULL, nrep = 100, weight = FALSE,
  mc.cores = 1)
}
\arguments{
\item{xy1}{Matrix or data frame of occurrence coordinates for species 1.}

\item{xy2}{Matrix or data frame of occurrence coordinates for species 2.}

\item{...}{Other matrices or data frames of occurrence coordinates for additional species.}

\item{sdm}{Raster or RasterStack representing environmental suitability (can be predictions from SDMs). 
These must have the same extent as both species' occurrence points. Default is NULL.}

\item{nrep}{Numeric for number of SVM tuning iterations. Default is 100.}

\item{weight}{Boolean. If TRUE, species with fewer occurrence records are weighted higher in the SVM. Default is FALSE.}

\item{mc.cores}{Number of cores to use for parallel processing. Default is 1.}
}
\value{
The tuned SVM model.
}
\description{
\code{rangeSVM()} returns a tuned support vector machine (SVM) model that predicts 
species identity based on predictors that are solely spatial, based on occurrence coordinates,
or a combination of spatial and environmental, based on both occurrence coordinates and 
environmental suitability values. Suitability values can be predicted with species distribution
models (SDMs; a.k.a. ecological niche models).
}
\details{
The tuning operation uses \code{tune.svm()} from the e1071 package, which performs 10-fold
cross validation and selects the best model based on classification error. Ranges of the cost and gamma
parameters are explored in the tuning exercise. The tuning function is iterated \code{nrep} times, and the
parameter combination used most frequently across all iterations is used to build a final SVM model.

When \code{sdm = NULL}, the SVM is purely spatial, based only on the occurrence coordinates of
each species. Otherwise, the SVM is fit with both a spatial predictor and any additional ones added as
rasters. These extra predictors can be based on predictions from a species distribution model 
(SDM; a.k.a. ecological niche model), and in this case would represent environmental or climatic
suitability, depending on the variables used in the SDM.
}
\examples{
\dontrun{
########## Maksing by biotic interactions
simulateData()
# Spatial SVMs
svm.SP <- rangeSVM(sp1.xy, sp2.xy, sp3.xy, nrep=10)
# Use SVM to create a raster of predicted regions
rand_svm.SP <- rangeSVM_predict(svm = svm.SP, r = r1.svm)
# Plot the results
plot(rand_svm.SP, col=c("yellow","pink","lightblue"))
points(sp1.xy, pch = 20, cex = 0.75, col = "orange")
points(sp2.xy, pch = 20, cex = 0.75, col = "green")
points(sp3.xy, pch = 20, cex = 0.75, col = "blue")
# We can use this SVM as a mask over our original SDM predictions.
# masked SDM predictions for variegatus
sp1_svm.SP <- rand_svm.SP == 1
sp1_svm.SP[sp1_svm.SP == 0] <- NA
sp1_svmSP_mask <- mask(r1.sdm, sp1_svm.SP)
# masked SDM predictions for tridactylus
sp2_svm.SP <- rand_svm.SP == 2
sp2_svm.SP[sp2_svm.SP == 0] <- NA
sp2_svmSP_mask <- mask(r2.sdm, sp2_svm.SP)
# masked SDM predictions for torquatus
sp3_svm.SP <- rand_svm.SP == 3
sp3_svm.SP[sp3_svm.SP == 0] <- NA
sp3_svmSP_mask <- mask(r3.sdm, sp3_svm.SP)
##Plot the predicted realized distributions for each species.
# Create 3-panel figure
par(mfrow = c(1,3))
plot(sp1_svmSP_mask)
plot(sp2_svmSP_mask)
plot(sp3_svmSP_mask)
## Hybrid SVMs
# Create SVM
svmHYB <- rangeSVM(sp1.xy, sp2.xy, sp3.xy, 
                   sdm = raster::stack(r1.sdm, r2.sdm, r3.sdm), 
                   nrep = 10)
# Use SVM to create a raster of predicted regions
rand_svmHYB <- rangeSVM_predict(svm = svmHYB, 
                                r = r1.sdm, 
                                sdm = raster::stack(r1.sdm, r2.sdm, r3.sdm))
## Plot the SVM results.
plot(rand_svmHYB, col=c("yellow","pink","lightblue"))
points(sp1.xy, pch = 20, cex = 0.75, col = "orange")
points(sp2.xy, pch = 20, cex = 0.75, col = "green")
points(sp3.xy, pch = 20, cex = 0.75, col = "blue")
## Use the hybrid SVM as a mask over our SDM predictions.
# masked SDM predictions for variegatus
sp1_svmHYB <- rand_svmHYB == 1
sp1_svmHYB[rand_svmHYB == 0] <- NA
sp1_svmHYB_mask <- mask(r1.sdm, sp1_svmHYB)
# masked SDM predictions for tridactylus
sp2_svmHYB <- rand_svmHYB == 2
sp2_svmHYB[sp2_svmHYB == 0] <- NA
sp2_svmHYB_mask <- mask(r2.sdm, sp2_svmHYB)
# masked SDM predictions for torquatus
sp3_svmHYB <- rand_svmHYB == 3
sp3_svmHYB[sp3_svmHYB == 0] <- NA
sp3_svmHYB_mask <- mask(r3.sdm, sp3_svmHYB)
## Plot the predicted realized distributions for each species.
# Create 3-panel figure
par(mfrow = c(1,3))
plot(sp1_svmHYB_mask)
plot(sp2_svmHYB_mask)
plot(sp3_svmHYB_mask)
}
}
